refactor: move routing interfaces into their own files
