+++ src/PepperDash.Essentials.Core/PepperDash.Essentials.Core.csproj
@@ -22,8 +22,11 @@
     <DebugType>pdbonly</DebugType>
   </PropertyGroup>
   <ItemGroup>
-    <PackageReference Include="Crestron.SimplSharp.SDK.ProgramLibrary" Version="2.20.66" />
-    <PackageReference Include="PepperDashCore" Version="2.0.0-beta-415" />
+    <Compile Remove="Routing\GenericExtensions.cs" />
+  </ItemGroup>
+  <ItemGroup>
+    <PackageReference Include="Crestron.SimplSharp.SDK.ProgramLibrary" Version="2.20.42" />
+    <PackageReference Include="PepperDashCore" Version="2.0.0-alpha-402" />
   </ItemGroup>
   <ItemGroup>
     <None Include="Crestron\CrestronGenericBaseDevice.cs.orig" />
