+++ src/PepperDash.Essentials.Core/Routing/IRmcRoutingWithFeedback.cs
@@ -1,0 +1,11 @@
+﻿using PepperDash.Essentials.Core.Routing;
+
+namespace PepperDash.Essentials.Core
+{
+    /// <summary>
+    /// Defines an IRmcRouting with a feedback event 
+    /// </summary>
+    public interface IRmcRoutingWithFeedback : IRmcRouting
+    {
+    }
+}
