+++ src/PepperDash.Essentials.Core/Routing/IRoutingInputsOutputs.cs
@@ -1,0 +1,16 @@
+﻿namespace PepperDash.Essentials.Core
+{
+    /// <summary>
+    /// For devices like RMCs, baluns, other devices with no switching.
+    /// </summary>
+    public interface IRoutingInputsOutputs : IRoutingInputs, IRoutingOutputs
+	{
+	}
+
+/*    /// <summary>
+    /// For devices like RMCs, baluns, other devices with no switching.
+    /// </summary>
+    public interface IRoutingInputsOutputs<TInputSelector, TOutputSelector> : IRoutingInputs<TInputSelector>, IRoutingOutputs<TOutputSelector>
+    {
+    }*/
+}
