+++ src/PepperDash.Essentials.Core/Routing/IRoutingNumeric.cs
@@ -1,0 +1,7 @@
+﻿namespace PepperDash.Essentials.Core
+{
+    public interface IRoutingNumeric : IRouting
+    {
+        void ExecuteNumericSwitch(ushort input, ushort output, eRoutingSignalType type);
+    }
+}
