+++ src/PepperDash.Essentials.Core/Routing/IRoutingNumericWithFeedback.cs
@@ -1,0 +1,9 @@
+﻿namespace PepperDash.Essentials.Core
+{
+    /// <summary>
+    /// Defines an IRoutingNumeric with a feedback event 
+    /// </summary>
+    public interface IRoutingNumericWithFeedback : IRoutingNumeric, IRoutingFeedback
+    {
+    }
+}
