+++ src/PepperDash.Essentials.Core/Routing/IRoutingSource.cs
@@ -1,0 +1,9 @@
+﻿namespace PepperDash.Essentials.Core
+{
+    /// <summary>
+    /// Defines an IRoutingOutputs devices as being a source - the start of the chain
+    /// </summary>
+    public interface IRoutingSource : IRoutingOutputs
+	{
+	}
+}
