+++ src/PepperDash.Essentials.Core/Routing/ITxRouting.cs
@@ -1,0 +1,8 @@
+﻿namespace PepperDash.Essentials.Core
+{
+    public interface ITxRouting : IRoutingNumeric
+    {
+        IntFeedback VideoSourceNumericFeedback { get; }
+        IntFeedback AudioSourceNumericFeedback { get; }
+    }
+}
