+++ src/PepperDash.Essentials.Core/Routing/ITxRoutingWithFeedback.cs
@@ -1,0 +1,9 @@
+﻿namespace PepperDash.Essentials.Core
+{
+    /// <summary>
+    /// Defines an IRmcRouting with a feedback event 
+    /// </summary>
+    public interface ITxRoutingWithFeedback : ITxRouting
+    {
+    }
+}
