+++ src/PepperDash.Essentials.Core/Routing/RoutingPortCollection.cs
@@ -23,4 +23,21 @@ namespace PepperDash.Essentials.Core
 			}
 		}
 	}
+
+/*    /// <summary>
+	/// Basically a List , with an indexer to find ports by key name
+	/// </summary>
+	public class RoutingPortCollection<T, TSelector> : List<T> where T : RoutingPort<TSelector>
+    {
+        /// <summary>
+        /// Case-insensitive port lookup linked to ports' keys
+        /// </summary>
+        public T this[string key]
+        {
+            get
+            {
+                return this.FirstOrDefault(i => i.Key.Equals(key, StringComparison.OrdinalIgnoreCase));
+            }
+        }
+    }*/
 }
