+++ src/PepperDash.Essentials.Core/Routing/eRoutingPortConnectionType.cs
@@ -1,0 +1,8 @@
+﻿namespace PepperDash.Essentials.Core
+{
+    public enum eRoutingPortConnectionType
+	{
+		None, BackplaneOnly, DisplayPort, Dvi, Hdmi, Rgb, Vga, LineAudio, DigitalAudio, Sdi, 
+		Composite, Component, DmCat, DmMmFiber, DmSmFiber, Speaker, Streaming, UsbC, HdBaseT
+	}
+}
