+++ src/PepperDash.Essentials.Core/Routing/eRoutingSignalType.cs
@@ -1,0 +1,16 @@
+﻿using System;
+
+
+namespace PepperDash.Essentials.Core
+{
+    [Flags]
+	public enum eRoutingSignalType
+	{
+		Audio = 1,
+		Video = 2,     
+		AudioVideo = Audio | Video,
+        UsbOutput = 8,
+        UsbInput = 16,
+        SecondaryAudio = 32
+	}
+}
